+++ src/gamspy/_algebra/expression.py
@@ -586,13 +586,13 @@ class Expression(operable.Operable):
 
         return records["value"][0]
 
-    def toList(self) -> list | None:
+    def toList(self) -> list:
         """
         Convenience method to return the records of the expression as a list.
 
         Returns
         -------
-        list | None
+        list
 
         Examples
         --------
@@ -610,7 +610,7 @@ class Expression(operable.Operable):
         if records is not None:
             return records.values.tolist()
 
-        return None
+        return []
 
     def __eq__(self, other):
         return Expression(self, "=e=", other)